requirements.txt
fastapi
uvicorn
pydantic
numpy
pandas
scikit-learn
xgboost
sqlalchemy
python-dotenv
joblib
groq